+++ packages/react-leaflet/src/ImageOverlay.tsx
@@ -5,7 +5,7 @@ import {
   extendContext,
   updateMediaOverlay,
 } from '@react-leaflet/core'
-import { ImageOverlay as LeafletImageOverlay } from 'leaflet'
+import { LatLngBounds, ImageOverlay as LeafletImageOverlay } from 'leaflet'
 import type { ReactNode } from 'react'
 
 export interface ImageOverlayProps extends MediaOverlayProps {
@@ -25,6 +25,13 @@ export const ImageOverlay = createLayerComponent<
   },
   function updateImageOverlay(overlay, props, prevProps) {
     updateMediaOverlay(overlay, props, prevProps)
+    if (props.bounds !== props.bounds) {
+      const bounds =
+        props.bounds instanceof LatLngBounds
+          ? props.bounds
+          : new LatLngBounds(props.bounds)
+      overlay.setBounds(bounds)
+    }
     if (props.url !== prevProps.url) {
       overlay.setUrl(props.url)
     }